pg flags, test of gauss tb, float W

+++ libnmsim/nmsim_elem_synapse.c
@@ -1,5 +1,5 @@
 /* See {nmsim_elem_synapse.h} */
-/* Last edited on 2020-12-06 16:08:34 by jstolfi */
+/* Last edited on 2020-12-09 22:42:39 by jstolfi */
 
 #define _GNU_SOURCE
 #include <stdint.h>
@@ -23,7 +23,7 @@ void nmsim_elem_synapse_write(FILE *wr, nmsim_elem_synapse_ix_t ise, nmsim_elem_
   {
     char *ind1 = "  ";   /* Indent of synapse line. */
     fprintf(wr, "%s%d %d  %d %d  ", ind1, ise, syn->isg, syn->ine_pre, syn->ine_pos);
-    nmsim_write_double_value(wr, syn->W, nmsim_write_KW_PREC, TRUE, TRUE, FALSE);
+    nmsim_write_double_value(wr, (double)syn->W, nmsim_write_KW_PREC, TRUE, TRUE, FALSE);
   }
 
 nmsim_elem_synapse_t nmsim_elem_synapse_read
@@ -40,7 +40,7 @@ nmsim_elem_synapse_t nmsim_elem_synapse_read
       (nmsim_elem_neuron_ix_t)nmsim_read_int64_value(rd, "pre-synaptic neuron index", 0, ine_max);
     nmsim_elem_neuron_ix_t ine_pos = 
       (nmsim_elem_neuron_ix_t)nmsim_read_int64_value(rd, "post-synaptic neuron index", 0, ine_max);
-    double W = nmsim_read_double_value(rd, "resting synaptic weight", -1000.0, +1000.0);
+    float W = (float)nmsim_read_double_value(rd, "resting synaptic weight", -1000.0, +1000.0);
     
     nmsim_elem_synapse_t syn = (nmsim_elem_synapse_t)
       { .isg = isg, .ine_pre = ine_pre, .ine_pos = ine_pos, .W = W };
@@ -81,11 +81,11 @@ nmsim_elem_synapse_t nmsim_elem_synapse_throw
     nmsim_elem_neuron_ix_t ine_pos = (nmsim_elem_neuron_ix_t)int64_abrandom(ine_pos_min, ine_pos_max);
     
     /* Pick a synaptic strength {W}: */
-    double W = dloggaussrand(W_avg, W_dev);
+    float W = (float)dloggaussrand(W_avg, W_dev);
     if (W_dev != 0.0) 
       { fprintf(stderr, "  synapse weight %18.12f", W);
         double mod = nmsim_select_rounding_mod(-W_dev/100, +W_dev/100);
-        W = mod*floor(W/mod + 0.5);
+        W = (float)(mod*floor(W/mod + 0.5));
         fprintf(stderr, " rounding modulus = %18.12f rounded to %18.12f\n", mod, W);
       }
